backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
httpx>=0.27,<1.0
beautifulsoup4>=4.12,<5.0
pypdf>=5.0,<6.0
python-dotenv>=1.0,<2.0
langchain-core>=0.3,<1.0
langgraph>=0.2,<1.0
langsmith>=0.1,<1.0
pinecone>=5.0,<6.0
pytest>=8.3,<9.0
pytest-asyncio>=0.24,<1.0
ragas>=0.2,<1.0
datasets>=3.0,<4.0
langchain-google-genai>=2.0,<3.0
